REFACTOR: change document to model module in core layer
